Reorganzing Controllers
Since Friday we pretty much listened to how our app is organized wrong
and the controllers aren't meant to do any kind of logic based stuff I'm
just moving some of the functionality into the model classes, or trying
to anyway. Imma keeyp this in a branch for now though in case I break
stuff badly.